+++ R/mod_review_form_tbl.R
@@ -27,6 +27,12 @@ mod_review_form_tbl_ui <- function(id) {
 #' @param form_review_data Common reactive value containing the review data of
 #'   the form.
 #' @param form_items Named character vector with all form_items to display.
+#' @param transformation A reactive value. If this is 'none' then the columns
+#'   `item_value` and `item_unit` will be used in the table. Otherwise,
+#'   `value_standardized` and `unit_standardized` will be used.
+#' @param show_limits Optional reactive value containing a logical. If the
+#'   logical inside is `TRUE`, laboratory limits will be added to the table
+#'   shown in the module.
 #' @param active_subject Reactive value containing the active subject id.
 #' @param show_all Common reactive value, a logical indicating whether all
 #'   records should be displayed.
@@ -35,7 +41,7 @@ mod_review_form_tbl_ui <- function(id) {
 #'   interactive tables.
 #' @param title An optional character vector. If provided, will be used within
 #'   [datatable_custom()], as the title for the table.
-#'   
+#'
 #' @seealso [mod_review_form_tbl_ui()], [mod_common_forms_ui()],
 #'   [mod_common_forms_server()], [mod_study_forms_ui()],
 #'   [mod_study_forms_server()]
@@ -46,6 +52,8 @@ mod_review_form_tbl_server <- function(
     form_data,
     form_review_data,
     form_items,
+    transformation = NULL,
+    show_limits = NULL,
     active_subject, 
     show_all,
     table_names = NULL,
@@ -59,6 +67,8 @@ mod_review_form_tbl_server <- function(
   stopifnot(is.reactive(show_all))
   stopifnot(is.character(table_names %||% ""))
   stopifnot(is.character(title %||% ""))
+  transformation <- transformation %||% reactiveVal("none")
+  show_limits <- show_limits %||% reactiveVal(FALSE)
 
   moduleServer(id, function(input, output, session){
     ns <- session$ns
@@ -78,12 +88,14 @@ mod_review_form_tbl_server <- function(
         form_review_data(), 
         form = form, 
         form_items = form_items,
+        transformation = transformation() %||% "none",
+        show_limits = show_limits() %||% FALSE,
         active_subject = if(identical(session$userData$review_type(), "form")) NULL else active_subject(),
         pending_form_review_status = NULL,
         is_SAE = identical(title, "Serious Adverse Events")
       )
     }) |> 
-      bindEvent(form_data(), form_review_data(), active_subject(), session$userData$review_type())
+      bindEvent(form_data(), form_review_data(), active_subject(), session$userData$review_type(), transformation(), show_limits())
     
     ############################### Observers: #################################
     
@@ -94,7 +106,7 @@ mod_review_form_tbl_server <- function(
       session$userData$pending_form_review_status[[form]] <- NULL
       session$userData$pending_review_records[[form]] <- data.frame(id = integer(), reviewed = character())
     }, priority = 100) |> 
-      bindEvent(active_subject(), form_review_data(), form_data(), session$userData$review_type())
+      bindEvent(active_subject(), form_review_data(), form_data(), session$userData$review_type(), transformation(), show_limits())
     
     observeEvent(datatable_rendered(), {
       golem::cat_dev(form, "| renewing table_data using merged_form_data()\n\n")