+++ NEWS.md
@@ -7,7 +7,10 @@
 - Added a toggle to enable hover labels of all patterns in the figures, also the ones in the background.
 - Improved the time-series on-hover label so that limits and significance are also shown.
 - Switched to server-side table downloads.
-- Implemented a `line-clamp` on table column names that wrap to more than 6 lines
+- Implemented a `line-clamp` on table column names that wrap to more than 6 lines.
+- Added a toggle to switch data transformation. In addition, added a way to show data transformed to standard values so that all are shown in the same unit (#110, #261).
+- Figures now also show patient-specific lab limits in the figures when data is not transformed (if available).
+- Added additional navigation buttons next to the tabs to easier navigate between common forms and study forms (#255).
 
 ## Bug fixes 
 
@@ -16,6 +19,7 @@
 ## Developer notes
 
 - Updated role of long-term contributors to co-authors in Description field (#246).
+- Refactored some functions so that it is no longer needed to create an `apptables` object when starting the applcation, improving start up efficiency (#251).
 
 # clinsight 0.3.0